Allow collection queries without key references
For both BTree and HashMap, it is possible to have keys whose full value
is not used in the equality check.  It is sometimes useful to look up
items in the map using a newtype that can hash or compare itself with
actual keys, without being a valid key itself.

This extends (some methods of) these collections to allow more flexible
queries.  Currently an RFC.

Signed-off-by: Daniel De Graaf <code@danieldg.net>